requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
google-generativeai==0.3.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
transformers==4.36.0
torch
nltk==3.8.1
textstat==0.7.3
requests==2.31.0
easyocr
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1